Added Complete TestData Framework
Added the DataController, the DataConfig startup file, the DatabaseFile, and the Datamodel.

We can add/change all the parts to support whatever data we need. Right now the Database just has a super basic User table with ID and Name.

Note the DataConfig class is Static, so you dont need to instantiate it. And there's only one copy of the database interface per run-time execution.